refactor: extract desktop notifier into dedicated file
Move DesktopNotifier struct, Notify method, and encodeUTF16LEBase64
helper from notify.go into desktop.go. notify.go now contains only
the Notifier interface and New factory function.
